[Feat] 인기 게시글 조회 API 구현

+++ src/main/java/org/sopt/bofit/domain/post/dto/response/PostDetailResponse.java
@@ -1,10 +1,9 @@
 package org.sopt.bofit.domain.post.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.Builder;
-
 import java.time.LocalDateTime;
 import java.util.List;
+import lombok.Builder;
 
 @Builder
 public record PostDetailResponse(
@@ -24,13 +23,13 @@ public record PostDetailResponse(
         String content,
 
         @Schema(description = "댓글 수 ", example = "8")
-        long commentCount,
+        int commentCount,
 
         @Schema(description = "생성 시간")
         LocalDateTime createdAt,
 
         @Schema(description = "좋아요 수")
-        Long likeCount,
+        int likeCount,
 
         @Schema(description = "사용자 좋아요 여부")
         boolean likedByCurrentUser,
